Improve performance of Stopword lookup
Ankusa::STOPWORDS is created once and then searched for every single
word that we are classifying.

Change it from an Array with O(n) average time complexity to a Set
(Hash-Table) with O(1) average time complexity.

In tests on 1000 emails (Single threaded) the run-time was reduced
from 6.4 seconds to 2.4 seconds.